[kvmtest]: improve kvmtest to parse options (#2732)
add several options for user to define the kvmtest behavior

+++ tests/kvmtest.sh
@@ -1,25 +1,86 @@
 #!/bin/bash -xe
 
-tbname=$1
-dut=$2
+usage() {
+    cat >&2 <<EOF
+Usage:
+  kvmtest.sh [-n] [-i inventory] [-t testbed_file] [-d SONIC_MGMT_DIR] tbname dut
+
+Description:
+  -n
+       Do not refresh DUT
+  -i inventory
+       inventory file (default: veos_vtb)
+  -t testbed_file
+       testbed file (default: vtestbed.csv)
+  -d SONIC_MGMT_DIR
+       sonic-mgmt repo directory (default: /data/sonic-mgmt)
+  tbname
+       testbed name
+  dut
+       DUT name
+
+Example:
+  ./kvmtest.sh vms-kvm-t0 vlab-01
+EOF
+}
+
 inventory="veos_vtb"
 testbed_file="vtestbed.csv"
-
+refresh_dut=true
 SONIC_MGMT_DIR=/data/sonic-mgmt
 
-cd $HOME
-mkdir -p .ssh
-cp /data/pkey.txt .ssh/id_rsa
-chmod 600 .ssh/id_rsa
+while getopts "ni:t:d:" opt; do
+  case $opt in
+    n)
+      refresh_dut=""
+      ;;
+    i)
+      inventory=$OPTARG
+      ;;
+    t)
+      testbed_file=$OPTARG
+      ;;
+    d)
+      SONIC_MGMT_DIR=$OPTARG
+      ;;
+    \?)
+      echo "Invalid option: -$OPTARG" >&2
+      usage
+      exit 1
+      ;;
+  esac
+done
+shift "$((OPTIND - 1))"
 
-# Refresh dut in the virtual switch topology
-cd $SONIC_MGMT_DIR/ansible
-./testbed-cli.sh -m $inventory -t $testbed_file refresh-dut $tbname password.txt
-sleep 120
+if [ $# -lt 2 ]; then
+    usage
+    exit 1
+fi
+
+tbname=$1
+dut=$2
+
+
+if [ -f /data/pkey.txt ]; then
+    pushd $HOME
+    mkdir -p .ssh
+    cp /data/pkey.txt .ssh/id_rsa
+    chmod 600 .ssh/id_rsa
+    popd
+fi
+
+pushd $SONIC_MGMT_DIR/ansible
+if [ -n "$refresh_dut" ]; then
+    # Refresh dut in the virtual switch topology
+    ./testbed-cli.sh -m $inventory -t $testbed_file refresh-dut $tbname password.txt
+    sleep 120
+fi
 
 # Create and deploy default vlan configuration (one_vlan_a) to the virtual switch
 ./testbed-cli.sh -m $inventory -t $testbed_file deploy-mg $tbname lab password.txt
 sleep 180
+
+popd
 
 export ANSIBLE_LIBRARY=$SONIC_MGMT_DIR/ansible/library/
 
@@ -38,7 +99,7 @@ PYTEST_CLI_COMMON_OPTS="\
 -a False \
 -O"
 
-cd $SONIC_MGMT_DIR/tests
+pushd $SONIC_MGMT_DIR/tests
 rm -rf logs
 mkdir -p logs
 
@@ -78,13 +139,13 @@ iface_namingmode/test_iface_namingmode.py \
 platform_tests/test_cpu_memory_usage.py \
 bgp/test_bgpmon.py"
 
-pushd $SONIC_MGMT_DIR/tests
 ./run_tests.sh $PYTEST_CLI_COMMON_OPTS -c "$tests" -p logs/$tgname
 popd
 
 # Create and deploy two vlan configuration (two_vlan_a) to the virtual switch
-cd $SONIC_MGMT_DIR/ansible
+pushd $SONIC_MGMT_DIR/ansible
 ./testbed-cli.sh -m $inventory -t $testbed_file deploy-mg $tbname lab password.txt -e vlan_config=two_vlan_a
+popd
 sleep 180
 
 # Run tests_2vlans on vlab-01 virtual switch
